Conflicts:
	README.md

+++ container.go
@@ -3,8 +3,8 @@ package docker
 import (
 	"encoding/json"
 	"errors"
-	"flag"
 	"fmt"
+	"github.com/dotcloud/docker/rcli"
 	"github.com/kr/pty"
 	"io"
 	"io/ioutil"
@@ -60,9 +60,12 @@ type Config struct {
 	Image      string // Name of the image as it was passed by the operator (eg. could be symbolic)
 }
 
-func ParseRun(args []string) (*Config, error) {
-	cmd := flag.NewFlagSet("", flag.ContinueOnError)
-	cmd.SetOutput(ioutil.Discard)
+func ParseRun(args []string, stdout io.Writer) (*Config, error) {
+	cmd := rcli.Subcmd(stdout, "run", "[OPTIONS] IMAGE COMMAND [ARG...]", "Run a command in a new container")
+	if len(args) > 0 && args[0] != "--help" {
+		cmd.SetOutput(ioutil.Discard)
+	}
+
 	fl_user := cmd.String("u", "", "Username or UID")
 	fl_detach := cmd.Bool("d", false, "Detached mode: leave the container running in the background")
 	fl_stdin := cmd.Bool("i", false, "Keep stdin open even if not attached")
@@ -76,6 +79,15 @@ func ParseRun(args []string) (*Config, error) {
 	if err := cmd.Parse(args); err != nil {
 		return nil, err
 	}
+	parsedArgs := cmd.Args()
+	runCmd := []string{}
+	image := ""
+	if len(parsedArgs) >= 1 {
+		image = cmd.Arg(0)
+	}
+	if len(parsedArgs) > 1 {
+		runCmd = parsedArgs[1:]
+	}
 	config := &Config{
 		Ports:     fl_ports,
 		User:      *fl_user,
@@ -84,8 +96,8 @@ func ParseRun(args []string) (*Config, error) {
 		Memory:    *fl_memory,
 		Detach:    *fl_detach,
 		Env:       fl_env,
-		Cmd:       cmd.Args()[1:],
-		Image:     cmd.Arg(0),
+		Cmd:       runCmd,
+		Image:     image,
 	}
 	return config, nil
 }
@@ -246,6 +258,10 @@ func (container *Container) Start() error {
 
 	var err error
 	if container.Config.Tty {
+		container.cmd.Env = append(
+			[]string{"TERM=xterm"},
+			container.cmd.Env...,
+		)
 		err = container.startPty()
 	} else {
 		err = container.start()
@@ -356,6 +372,9 @@ func (container *Container) monitor() {
 }
 
 func (container *Container) kill() error {
+	if container.cmd == nil {
+		return nil
+	}
 	if err := container.cmd.Process.Kill(); err != nil {
 		return err
 	}
